-> Shuffling around methods for cohort viewing
