+++ cohorts/models.py
@@ -119,7 +119,7 @@ class Cohort(models.Model):
     def get_data_sources(self, source_type=None):
         result = {}
 
-        cohort_filters = Filters.objects.select_related('attribute').filter(resulting_cohort=self)
+        cohort_filters = Filter.objects.select_related('attribute').filter(resulting_cohort=self)
         attributes = Attribute.objects.filter(id__in=cohort_filters.values_list('attribute', flat=True))
 
         data_versions = self.get_data_versions()
@@ -139,7 +139,7 @@ class Cohort(models.Model):
     # Returns the set of filters defining this cohort as a dict organized by data source
     def get_filters_by_data_source(self, source_type=None):
 
-        cohort_filters = Filters.objects.select_related('attribute').filter(resulting_cohort=self)
+        cohort_filters = Filter.objects.select_related('attribute').filter(resulting_cohort=self)
         result = self.get_data_sources(source_type)
 
         for source in DataSource.SOURCE_TYPES:
@@ -166,7 +166,7 @@ class Cohort(models.Model):
                 'data_versions': [x.name for x in dvs],
                 'filters': []
             }
-            filters = fg.filters_set.all()
+            filters = fg.filter_set.all()
             attributes = Attribute.objects.filter(id__in=filters.values_list('attribute', flat=True))
             for attr in attributes:
                 group['filters'].append({
