+++ cohorts/views/views.py
@@ -53,7 +53,8 @@ from django.utils.html import escape
 
 from cohorts.models import Cohort, Cohort_Perms, Source, Filter, Filter_Group, Cohort_Comments
 from cohorts.utils import _save_cohort, _delete_cohort, get_cohort_uuids
-from idc_collections.models import Program, Collection, DataSetType, DataVersion
+from idc_collections.models import Program, Collection, DataSource, DataVersion
+from idc_collections.collex_metadata_utils import build_explorer_context
 
 BQ_ATTEMPT_MAX = 10
 
@@ -139,242 +140,6 @@ def cohorts_list(request, is_public=False):
                                                         'previously_selected_cohort_ids' : previously_selected_cohort_ids
                                                         })
 
-
-# Build data exploration context/response
-def build_explorer_context(is_dicofdic, source, versions, filters, fields, order_docs, counts_only, with_related,
-                           with_derived, collapse_on, is_json):
-    attr_by_source = {}
-    attr_sets = {}
-    context = {}
-
-    try:
-        context['collection_tooltips'] = Collection.objects.filter(active=True).get_tooltips()
-
-        versions = DataVersion.objects.filter(name__in=versions) if len(versions) else DataVersion.objects.filter(
-            active=True)
-
-        data_types = [DataSetType.IMAGE_DATA, ]
-        with_related and data_types.extend(DataSetType.ANCILLARY_DATA)
-        with_derived and data_types.extend(DataSetType.DERIVED_DATA)
-        data_sets = DataSetType.objects.filter(data_type__in=data_types)
-        sources = data_sets.get_data_sources().filter(source_type=source, id__in=versions.get_data_sources().filter(
-            source_type=source).values_list("id", flat=True)).distinct()
-
-        source_attrs = sources.get_source_attrs(for_ui=True, with_set_map=True)
-
-        # For now we're only allowing TCGA+ispy1+lidc-idri+qin_headneck
-        # TODO: REMOVE THIS ONCE WE'RE ALLOWING MORE
-        tcga_in_tcia = Program.objects.get(short_name="TCGA").collection_set.all()
-        collectionFilterList = [collex.collection_id for collex in tcga_in_tcia] + ['ispy1', 'lidc_idri',
-                                                                                    'qin_headneck', 'nsclc_radiomics']
-        if not 'collection_id' in filters:
-            filters['collection_id'] = collectionFilterList
-
-        source_data_types = sources.get_source_data_types()
-
-        for source in sources:
-            is_origin = DataSetType.IMAGE_DATA in source_data_types[source.id]
-            # If a field list wasn't provided, work from a default set
-            if is_origin and not len(fields):
-                fields = source.get_attr(for_faceting=False).filter(default_ui_display=True).values_list('name',
-                                                                                                         flat=True)
-
-            for dataset in data_sets:
-                if dataset.data_type in source_data_types[source.id]:
-                    set_type = dataset.get_set_name()
-                    if set_type not in attr_by_source:
-                        attr_by_source[set_type] = {}
-                    attrs = source_attrs['sources'][source.id]['attr_sets'][dataset.id]
-                    if 'attributes' not in attr_by_source[set_type]:
-                        attr_by_source[set_type]['attributes'] = {}
-                        attr_sets[set_type] = attrs
-                    else:
-                        attr_sets[set_type] = attr_sets[set_type] | attrs
-
-                    attr_by_source[set_type]['attributes'].update(
-                        {attr.name: {'source': source.id, 'obj': attr, 'vals': None, 'id': attr.id} for attr in attrs}
-                    )
-
-        start = time.time()
-        source_metadata = get_collex_metadata(
-            filters, fields, record_limit=1000, counts_only=counts_only, with_ancillary=with_related,
-            collapse_on=collapse_on, order_docs=order_docs, sources=sources, versions=versions
-        )
-        stop = time.time()
-        logger.debug("[STATUS] Benchmarking: Time to collect metadata for source type {}: {}s".format(
-            "BigQuery" if sources.first().source_type == DataSource.BIGQUERY else "Solr",
-            str((stop - start))
-        ))
-
-        for source in source_metadata['facets']:
-            source_name = ":".join(source.split(":")[0:2])
-            facet_set = source_metadata['facets'][source]['facets']
-            for dataset in data_sets:
-                if dataset.data_type in source_data_types[int(source.split(":")[-1])]:
-                    set_name = dataset.get_set_name()
-                    if dataset.data_type in data_types and set_name in attr_sets:
-                        attr_display_vals = Attribute_Display_Values.objects.filter(
-                            attribute__id__in=attr_sets[set_name]).to_dict()
-                        if dataset.data_type == DataSetType.DERIVED_DATA:
-                            attr_cats = attr_sets[set_name].get_attr_cats()
-                            for attr in facet_set:
-                                if attr in attr_by_source[set_name]['attributes']:
-                                    source_name = "{}:{}".format(source_name.split(":")[0], attr_cats[attr]['cat_name'])
-                                    if source_name not in attr_by_source[set_name]:
-                                        attr_by_source[set_name][source_name] = {'attributes': {}}
-                                    attr_by_source[set_name][source_name]['attributes'][attr] = \
-                                        attr_by_source[set_name]['attributes'][attr]
-                                    this_attr = attr_by_source[set_name]['attributes'][attr]['obj']
-                                    values = []
-                                    for val in facet_set[attr]:
-                                        if val == 'min_max':
-                                            attr_by_source[set_name][source_name]['attributes'][attr][val] = \
-                                            facet_set[attr][val]
-                                        else:
-                                            displ_val = val if this_attr.preformatted_values else attr_display_vals.get(
-                                                this_attr.id, {}).get(val, None)
-                                            values.append({
-                                                'value': val,
-                                                'display_value': displ_val,
-                                                'units': this_attr.units,
-                                                'count': facet_set[attr][val] if val in facet_set[attr] else 0
-                                            })
-                                    attr_by_source[set_name][source_name]['attributes'][attr]['vals'] = sorted(values,
-                                                                                                               key=lambda
-                                                                                                                   x: x[
-                                                                                                                   'value'])
-                        else:
-                            attr_by_source[set_name]['All'] = {'attributes': attr_by_source[set_name]['attributes']}
-                            for attr in facet_set:
-                                if attr in attr_by_source[set_name]['attributes']:
-                                    this_attr = attr_by_source[set_name]['attributes'][attr]['obj']
-                                    values = []
-                                    for val in source_metadata['facets'][source]['facets'][attr]:
-                                        if val == 'min_max':
-                                            attr_by_source[set_name]['All']['attributes'][attr][val] = facet_set[attr][
-                                                val]
-                                        else:
-                                            displ_val = val if this_attr.preformatted_values else attr_display_vals.get(
-                                                this_attr.id, {}).get(val, None)
-                                            values.append({
-                                                'value': val,
-                                                'display_value': displ_val,
-                                                'count': facet_set[attr][val] if val in facet_set[attr] else 0
-                                            })
-                                    if attr == 'bmi':
-                                        sortDic = {'underweight': 0, 'normal weight': 1, 'overweight': 2, 'obese': 3,
-                                                   'None': 4}
-                                        attr_by_source[set_name]['All']['attributes'][attr]['vals'] = sorted(values,
-                                                                                                             key=lambda
-                                                                                                                 x:
-                                                                                                             sortDic[x[
-                                                                                                                 'value']])
-                                    else:
-                                        attr_by_source[set_name]['All']['attributes'][attr]['vals'] = sorted(values,
-                                                                                                             key=lambda
-                                                                                                                 x: x[
-                                                                                                                 'value'])
-
-        for set in attr_by_source:
-            for source in attr_by_source[set]:
-                if source == 'attributes':
-                    continue
-                if is_dicofdic:
-                    for x in list(attr_by_source[set][source]['attributes'].keys()):
-                        if (isinstance(attr_by_source[set][source]['attributes'][x]['vals'], list) and (
-                                len(attr_by_source[set][source]['attributes'][x]['vals']) > 0)):
-                            attr_by_source[set][source]['attributes'][x] = {y['value']: {
-                                'display_value': y['display_value'], 'count': y['count']
-                            } for y in attr_by_source[set][source]['attributes'][x]['vals']}
-                        else:
-                            attr_by_source[set][source]['attributes'][x] = {}
-
-                    if set == 'origin_set':
-                        context['collections'] = {
-                        a: attr_by_source[set][source]['attributes']['collection_id'][a]['count'] for a in
-                        attr_by_source[set][source]['attributes']['collection_id']}
-                        context['collections']['All'] = source_metadata['total']
-                else:
-                    if set == 'origin_set':
-                        collex = attr_by_source[set][source]['attributes']['collection_id']
-                        if collex['vals']:
-                            context['collections'] = {a['value']: a['count'] for a in collex['vals'] if
-                                                      a['value'] in collectionFilterList}
-                        else:
-                            context['collections'] = {a.name: 0 for a in Collection.objects.filter(active=True,
-                                                                                                   name__in=collectionFilterList)}
-                        context['collections']['All'] = source_metadata['total']
-
-                    attr_by_source[set][source]['attributes'] = [{
-                        'name': x,
-                        'id': attr_by_source[set][source]['attributes'][x]['obj'].id,
-                        'display_name': attr_by_source[set][source]['attributes'][x]['obj'].display_name,
-                        'values': attr_by_source[set][source]['attributes'][x]['vals'],
-                        'units': attr_by_source[set][source]['attributes'][x]['obj'].units,
-                        'min_max': attr_by_source[set][source]['attributes'][x].get('min_max', None)
-                    } for x, val in sorted(attr_by_source[set][source]['attributes'].items())]
-
-            if not counts_only:
-                attr_by_source[set]['docs'] = source_metadata['docs']
-
-        for key, source_set in attr_by_source.items():
-            sources = list(source_set.keys())
-            for key in sources:
-                if key == 'attributes':
-                    source_set.pop(key)
-
-        attr_by_source['total'] = source_metadata['total']
-
-        context['set_attributes'] = attr_by_source
-        context['filters'] = filters
-
-        programs = [x.lower() for x in list(Program.get_public_programs().values_list('short_name', flat=True))]
-        programSet = {}
-        for collection in context['collections']:
-            pref = collection.split('_')[0]
-            if pref in programs:
-                if not pref in programSet:
-                    programSet[pref] = {
-                        'projects': {},
-                        'val': 0
-                    }
-                programSet[pref]['projects'][collection] = context['collections'][collection]
-                programSet[pref]['val'] += context['collections'][collection]
-            else:
-                programSet[collection] = {'val': context['collections'][collection]}
-
-        if with_related:
-            context['tcga_collections'] = tcga_in_tcia
-
-        context['programs'] = programSet
-        # context['derived_list'] = [{'segmentations:TCIA Segmentation Analysis':'Segmentation'}, {'qualitative_measurements:TCIA Qualitative Analysis': 'Qualitative Analysis'}, {'quantitative_measurements:TCIA Quantitative Analysis':'Quantitative Analysis'}]
-
-        if 'derived_set' in context['set_attributes']:
-            if 'dicom_derived_all:segmentation' in context['set_attributes']['derived_set']:
-                context['set_attributes']['derived_set']['dicom_derived_all:segmentation'].update(
-                    {'display_name': 'Segmentation', 'name': 'segmentation'})
-            if 'dicom_derived_all:qualitative' in context['set_attributes']['derived_set']:
-                context['set_attributes']['derived_set']['dicom_derived_all:qualitative'].update(
-                    {'display_name': 'Qualitative Analysis', 'name': 'qualitative'})
-            if 'dicom_derived_all:quantitative' in context['set_attributes']['derived_set']:
-                context['set_attributes']['derived_set']['dicom_derived_all:quantitative'].update(
-                    {'display_name': 'Quantitative Analysis', 'name': 'quantitative'})
-
-        if is_json:
-            attr_by_source['programs'] = programSet
-            return attr_by_source
-        else:
-            context['order'] = {}
-            context['order']['derived_set'] = ['dicom_derived_all:segmentation', 'dicom_derived_all:qualitative',
-                                               'dicom_derived_all:quantitative']
-        return context
-
-    except Exception as e:
-        logger.error("[ERROR] While attempting to load the search page:")
-        logger.exception(e)
-
-    return None
-
 @login_required
 def cohort_detail(request, cohort_id):
     if debug: logger.debug('Called {}'.format(sys._getframe().f_code.co_name))
@@ -383,8 +148,6 @@ def cohort_detail(request, cohort_id):
         req = request.GET if request.GET else request.POST
         is_dicofdic = (req.get('is_dicofdic', "False").lower() == "true")
         source = req.get('data_source_type', DataSource.SOLR)
-        versions = json.loads(req.get('versions', '[]'))
-        filters = json.loads(req.get('filters', '{}'))
         fields = json.loads(req.get('fields', '[]'))
         order_docs = json.loads(req.get('order_docs', '[]'))
         counts_only = (req.get('counts_only', "False").lower() == "true")
@@ -404,20 +167,23 @@ def cohort_detail(request, cohort_id):
         shared_with_ids = Cohort_Perms.objects.filter(cohort=cohort, perm=Cohort_Perms.READER).values_list('user', flat=True)
         shared_with_users = User.objects.filter(id__in=shared_with_ids)
 
-        template_values = build_explorer_context(is_dicofdic, source, versions, filters, fields, order_docs, counts_only, with_related, with_derived, collapse_on, is_json)
+        cohort_filters = cohort.get_filters_as_dict()
+        cohort_versions = cohort.get_data_versions()
+        initial_filters = {}
+
+        template_values = build_explorer_context(is_dicofdic, source, cohort_versions, initial_filters, fields, order_docs, counts_only, with_related, with_derived, collapse_on, is_json)
 
         template_values.update({
             'request': request,
             'base_url': settings.BASE_URL,
-            'base_api_url': settings.BASE_API_URL
+            'cohort': cohort,
+            'shared_with_users': shared_with_users,
+            'export_url': reverse('export_data', kwargs={'cohort_id': cohort_id, 'export_type': 'cohort'}),
+            'cohort_filters': cohort_filters
         })
 
         template = 'cohorts/cohort_details.html'
-        template_values['cohort'] = cohort
-        template_values['total_samples'] = 0
-        template_values['total_cases'] = 0
-        template_values['shared_with_users'] = shared_with_users
-        template_values['export_url'] = reverse('export_data', kwargs={'cohort_id': cohort_id, 'export_type': 'cohort'})
+
 
     except ObjectDoesNotExist as e:
         logger.exception(e)
@@ -621,13 +387,13 @@ def clone_cohort(request, cohort_id):
         logger.debug('[BENCHMARKING] Time to builk create: ' + str(bulk_stop - bulk_start))
 
         # Clone the filters
-        filters = Filters.objects.filter(resulting_cohort=parent_cohort)
+        filters = Filter.objects.filter(resulting_cohort=parent_cohort)
         # ...but only if there are any (there may not be)
         if filters.__len__() > 0:
             filters_list = []
             for filter_pair in filters:
                 filters_list.append(Filters(name=filter_pair.name, value=filter_pair.value, resulting_cohort=cohort, program=filter_pair.program))
-            Filters.objects.bulk_create(filters_list)
+            Filter.objects.bulk_create(filters_list)
 
         # Set source
         source = Source(parent=parent_cohort, cohort=cohort, type=Source.CLONE)
